fix(templates/static): add logic to filter the color of team icons in the player cards
This change fixes a bug where team icons would not be displayed because they would have the same color as the card due to API's logic. Now, a filter alters the team-icon color to the teams' secondary color or adjust the contrast and saturation so the team icon appears properly.