feat(view): adjust logic to the see_spoilers session variable, fix the code to get all matches of a team to work with Postgres
Adjust the logic to the see_spoilers variable to create the variable in case it doesn't exist when the user accesses the home page. Additionally, add current segment name to the context of the home page. Also, adjust code to find all the matches of a team to work with Postgres. Finally, add (commented out) code to manually create or update the database.